remove references to the socorro package
A crude find and replace to reference the socorrolib library instead of socorro. Every package not in the current socorrolib will need to be moved there or integrated into this repository. The externals package requires some nuance to move.

This will not pass tests until there is a new socorrolib version with at least the above capabilities.